product digits
