+++ react-app/src/components/Products/ProductDetails.js
@@ -1,25 +1,16 @@
 import React, { useState } from 'react';
-import {  useDispatch, useSelector } from 'react-redux'
-import { useParams } from 'react-router-dom';
+import { Link, useParams } from 'react-router-dom';
 import './Product.css'
+// import {  useDispatch, useSelector } from 'react-redux'
 
 
-const ProductDetails = ({ productId }) =>{
+const ProductDetails = ({ product, userId, editMode, setEditMode }) =>{
     let editButton ;
 
-    const [editMode, setEditMode] = useState(false)
-
-    // const params = useParams()
-    const user = useSelector((state) => state.session.user)
-    let userId;
-    if (user) userId = user.id
-
-    // const productId = params.id
-    const product = useSelector((state) => state.products[productId])
-
     if ( userId === product.user_id) {
-        editButton = <button type='button' onClick={()=> setEditMode(true)}>Edit</button>
-
+        editButton = <Link to={`/products/${product.id}/edit`}>
+                        <button type='button' onClick={()=> setEditMode(true)}>Edit</button>
+                    </Link>
     }
 
     return (
@@ -31,7 +22,7 @@ const ProductDetails = ({ productId }) =>{
                 : <img src='https://i.imgur.com/BPOYKBx.png' alt={product.id}></img>}
             </div>
             <div className='product-description'>{product.description}</div>
-            <div className='product-price'>${product.price}</div>
+            <div className='product-price'>${product.price.toFixed(2)}</div>
             <div className='product-edit-btn'>{editButton}</div>
         </div>
     )
@@ -39,5 +30,15 @@ const ProductDetails = ({ productId }) =>{
 
 }
 
-
 export default ProductDetails
+
+
+// const [editMode, setEditMode] = useState(false)
+
+    // const params = useParams()
+    // const user = useSelector((state) => state.session.user)
+    // let userId;
+    // if (user) userId = user.id
+
+    // const productId = params.id
+    // const product = useSelector((state) => state.products[productId])
