+++ react-app/src/context/Modal.css
@@ -32,7 +32,7 @@
     justify-content: center;
     align-items: center;
     margin: auto;
-    padding: auto;
+    padding: 20px;
 }
 
 @media (max-width: 500px) {
